+++ src/devices/schneider_electric.ts
@@ -751,9 +751,7 @@ export const definitions: DefinitionWithExtend[] = [
                 powerOnBehavior: false,
                 color: false,
                 configureReporting: true,
-                levelConfig: {
-                    disabledFeatures: ["on_transition_time", "off_transition_time", "on_off_transition_time", "execute_if_off"],
-                },
+                levelConfig: {features: ["on_level", "current_level_startup"]},
             }),
             m.lightingBallast(),
             schneiderElectricExtend.dimmingMode(),
@@ -1920,22 +1918,7 @@ export const definitions: DefinitionWithExtend[] = [
         vendor: "Schneider Electric",
         description: "Wiser AvatarOn 1G dimmer switch",
         extend: [
-            m.light({
-                effect: false,
-                color: false,
-                powerOnBehavior: false,
-                levelConfig: {
-                    disabledFeatures: [
-                        "on_off_transition_time",
-                        "on_transition_time",
-                        "off_transition_time",
-                        "execute_if_off",
-                        "on_level",
-                        "current_level_startup",
-                    ],
-                },
-                configureReporting: true,
-            }),
+            m.light({effect: false, color: false, powerOnBehavior: false, configureReporting: true}),
             schneiderElectricExtend.addVisaConfigurationCluster(Zcl.DataType.UINT8),
             schneiderElectricExtend.visaConfigIndicatorLuminanceLevel(),
             schneiderElectricExtend.visaConfigIndicatorColor(),
@@ -1954,16 +1937,6 @@ export const definitions: DefinitionWithExtend[] = [
                 effect: false,
                 color: false,
                 powerOnBehavior: false,
-                levelConfig: {
-                    disabledFeatures: [
-                        "on_off_transition_time",
-                        "on_transition_time",
-                        "off_transition_time",
-                        "execute_if_off",
-                        "on_level",
-                        "current_level_startup",
-                    ],
-                },
                 configureReporting: true,
             }),
             schneiderElectricExtend.addVisaConfigurationCluster(Zcl.DataType.ENUM8),
@@ -2094,9 +2067,7 @@ export const definitions: DefinitionWithExtend[] = [
                 powerOnBehavior: false,
                 color: false,
                 configureReporting: true,
-                levelConfig: {
-                    disabledFeatures: ["on_transition_time", "off_transition_time", "on_off_transition_time", "execute_if_off"],
-                },
+                levelConfig: {features: ["on_level", "current_level_startup"]},
             }),
             m.lightingBallast(),
             m.illuminance(),
@@ -2129,7 +2100,7 @@ export const definitions: DefinitionWithExtend[] = [
                 color: false,
                 configureReporting: true,
                 levelConfig: {
-                    disabledFeatures: ["on_off_transition_time", "on_transition_time", "off_transition_time", "execute_if_off"],
+                    features: ["on_level", "current_level_startup"],
                 },
             }),
             m.lightingBallast(),